query_api/requirements.txt
fastapi==0.103.1
loguru==0.7.0
pymongo==4.5.0
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
